openaisdk/example1/pyproject.toml
[project]
name = "example1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.38.0",
    "openai-agents>=0.3.0",
]
